Move OrcidOAuth2 next to where it is used
